refactor(champs): extract dossier champ related methods in to a concern
